\name{data_recoding}
\alias{data_recoding}
\title{Recoding the variables of your dataframe}

\description{
  A versatile function designed for decoding answers from questionnaires generated by SKEZIA. The function requires two dataframes: one representing the answers and another specifying the variables, including their types.
}

\usage{
  data_recoding(df, bbl, sep = ", ")
}

\arguments{
  \item{df}{
    The dataframe containing answers to the SKEZIA questionnaire.
  }
  \item{bbl}{
    The dataframe containing variables you intend to recode in \code{df}.
  }
  \item{sep}{
    A separator used for parsing labels in the variables dataframe (\code{bbl}). Default is \code{', '}.
  }
}

\details{
  This function processes the variables dataframe (\code{bbl}) to extract information about variable names, numbers, and values. It then applies recoding logic to the answers dataframe (\code{df}), producing a modified version of the dataframe.
}

\value{
  Returns a recoded version of the answers dataframe (\code{df}).
}

\references{
  % Add any relevant references or literature related to the methodology or concepts used in the function.
}

\author{
  Akram Hecini
}

\note{
  % Add any additional notes or important information about the function.
}

\seealso{
  % Add objects to See Also as \code{\link{help}}.
}

\examples{
  \dontrun{
    # Example usage:
    recoded_df <- data_recoding(df = your_answers_dataframe, bbl = your_variables_dataframe)
  }
}
