\name{description_wordcloud}
\alias{description_wordcloud}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
How to make a word cloud with SKEZI
}

\usage{
wordcloud_maker(data, docs,...., tm_map)
}

\arguments{
  \item{data}{A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.}
  \item{toSpace}{Remove special characters by replacing with a wite space, by default set to @, /, ~.}
  \item{removePunctuation}{Remove any punctuation.}
  \item{removeWords}{Remove stopwords from word list, by default "french".}
  \item{removeNumbers}{Remove numbers from list.}
  \item{content_transformer}{Change all letters to either upper case or lower case, by default set to tolower.}
  \item{set.seed}{Random generator, meaning that he placement of words is randomised.}



}










\author{
Clementine LE RIDOU
}

